Fix code review & user reported bug
