+++ components/ArticleCategories/AddCategoryDialog.js
@@ -8,17 +8,6 @@ import ArticleCategory from './ArticleCategory';
 import CategoryOption from './CategoryOption';
 import Hint from 'components/NewReplySection/ReplyForm/Hint';
 
-const ADD_CATEGORY = gql`
-  mutation AddCategoryToArticle($articleId: String!, $categoryId: String!) {
-    CreateArticleCategory(articleId: $articleId, categoryId: $categoryId) {
-      articleId
-      categoryId
-      ...ArticleCategoryData
-    }
-  }
-  ${ArticleCategory.fragments.ArticleCategoryData}
-`;
-
 const CategoryData = gql`
   fragment CategoryData on Category {
     id
@@ -39,36 +28,6 @@ function AddCategoryDialog({
   articleCategories,
   onClose = () => {},
 }) {
-  const [addCategory, { loading }] = useMutation(ADD_CATEGORY, {
-    update(
-      cache,
-      {
-        data: { CreateArticleCategory },
-      }
-    ) {
-      // Read & update Article instance
-      const id = dataIdFromObject({ __typename: 'Article', id: articleId });
-      const article = cache.readFragment({
-        id,
-        fragmentName: 'ArticleWithCategories',
-        fragment: ArticleCategory.fragments.ArticleWithCategories,
-      });
-
-      cache.writeFragment({
-        id: dataIdFromObject({ __typename: 'Article', id: articleId }),
-        fragmentName: 'ArticleWithCategories',
-        fragment: ArticleCategory.fragments.ArticleWithCategories,
-        data: {
-          ...article,
-          articleCategories: CreateArticleCategory,
-        },
-      });
-    },
-  });
-
-  const handleAdd = categoryId => {
-    addCategory({ variables: { articleId, categoryId } });
-  };
 
   const feedbackMap = articleCategories.reduce((map, ac) => {
     map[ac.categoryId] = {
@@ -97,10 +56,8 @@ function AddCategoryDialog({
               key={category.id}
               articleId={articleId}
               category={category}
-              onAdd={handleAdd}
               feedback={feedbackMap[category.id]}
               marked={!!feedbackMap[category.id]}
-              disabled={loading}
             />
           ))}
         </div>
